#4902 - delete image, fixed image selection, moved text consts
