+++ packages/ketcher-core/src/application/editor/actions/rasterImage.ts
@@ -1,6 +1,10 @@
 import { ReStruct } from 'application/render';
 import { Vec2 } from 'domain/entities';
-import { Action, RasterImageUpsert } from 'application/editor';
+import {
+  Action,
+  RasterImageDelete,
+  RasterImageUpsert,
+} from 'application/editor';
 import { RasterImage } from 'domain/entities/rasterImage';
 
 export function fromRasterImageCreation(
@@ -12,5 +16,11 @@ export function fromRasterImageCreation(
   const action = new Action();
   const rasterImage = new RasterImage(bitmap, center, halfSize);
   action.addOp(new RasterImageUpsert(rasterImage));
+  return action.perform(reStruct);
+}
+
+export function fromRasterImageDeletion(reStruct: ReStruct, id: number) {
+  const action = new Action();
+  action.addOp(new RasterImageDelete(id));
   return action.perform(reStruct);
 }
