+++ packages/ketcher-core/src/application/render/restruct/restruct.ts
@@ -42,6 +42,7 @@ import Visel from './visel';
 import util from '../util';
 import { ReRGroupAttachmentPoint } from './rergroupAttachmentPoint';
 import { ReRasterImage } from 'application/render/restruct/rerasterImage';
+import { RASTER_IMAGE_KEY } from 'domain/constants';
 
 class ReStruct {
   public static readonly maps = {
@@ -58,9 +59,7 @@ class ReStruct {
     reloops: ReLoop,
     simpleObjects: ReSimpleObject,
     texts: ReText,
-    // We cannot use const RASTER_IMAGE_KEY here because this is being executed before initializing the variable
-    // But it must contain the same value otherwise the code will stop working
-    rasterImages: ReRasterImage,
+    [RASTER_IMAGE_KEY]: ReRasterImage,
   } as const;
 
   public render: Render;
