+++ packages/ketcher-core/src/domain/constants/index.ts
@@ -2,3 +2,4 @@ export * from './elementColor';
 export * from './elements';
 export * from './element.types';
 export * from './generics';
+export * from './rasterImage';
