+++ packages/ketcher-core/src/domain/constants/rasterImage.ts
@@ -1,0 +1,2 @@
+export const RASTER_IMAGE_KEY = 'rasterImages';
+export const RASTER_IMAGE_SERIALIZE_KEY = 'rasterImage';
